* inline menu editing (from within the game - to create new menus and to place nodes) * menu print function to show all values * changed align in menuNode_t from byte to int

+++ src/client/client.h
@@ -895,7 +895,7 @@ typedef struct menuNode_s {
 	vec2_t pos, size, texh, texl;
 	menuModel_t *menuModel;
 	byte state;
-	byte align;
+	int align;
 	byte invis, blend;
 	int mousefx;
 	int textScroll;				/* textfields - current scroll position */
